+++ R/panglm-package.R
@@ -3,8 +3,5 @@
 #' @keywords internal
 #' @useDynLib panglm, .registration = TRUE
 #' @importFrom Rcpp sourceCpp
-#' @importFrom data.table as.data.table setDT setkeyv rleidv
 #' @importFrom stats model.frame model.matrix model.response pnorm printCoefmat nobs
 "_PACKAGE"
-
-utils::globalVariables(c(".id", ".time", ".SD", "..cols", "..xnames", ":="))